+++ lumberjack.py
@@ -13,6 +13,8 @@ batch_buffer = []
 last_write_time = datetime.now()
 xp_buffer = {}
 trivia_questions = {}
+newquestion = {}
+questions_to_save = {}
 
 # Ensure the logs directory exists
 os.makedirs('logs', exist_ok=True)
@@ -39,9 +41,8 @@ conn.commit()
 conn.close()
 
 def flush_to_db():
-    """Flush the batch buffer to the database."""
     print('Flushing to database')
-    global batch_buffer, last_write_time, xp_buffer
+    global batch_buffer, last_write_time, xp_buffer, questions_to_save
     if not batch_buffer:
         return
 
@@ -52,6 +53,11 @@ def flush_to_db():
     #TODO: reduce writes by only updating xp if it has changed
     for user, xp in xp_buffer.items():
         c.execute("INSERT OR REPLACE INTO xp VALUES (?,?)", (user, xp))
+    #save trivia questions
+    for question, answer in questions_to_save.items():
+        c.execute("INSERT OR REPLACE INTO trivia_questions VALUES (?,?)", (question, answer))
+    #reset questions_to_save
+    questions_to_save = {}
     conn.commit()
     conn.close()